src/requirements.txt
args
arrow
asgiref
asn1crypto
async-timeout
attrs
autobahn
Automat
Babel
beautifulsoup4
cairocffi
CairoSVG
certifi
cffi
chardet
clint
constantly
cryptography
cssselect2
defusedxml
dj-database-url
Django
django-admin-honeypot
django-allauth
django-appconf
django-channels
django-ckeditor
django-compressor
django-crispy-forms
django-debug-toolbar
django-encrypted-model-fields
django-filter
django-haystack
django-hosts
django-js-asset
django-markdown
django-phonenumber-field
django-redis
django-request
django-softdelete
django-tables2
django-taggit
django-urlcrypt
django-weasyprint
django-widget-tweaks
djangorestframework
easy-thumbnails
elasticsearch
googledrivedownloader
holidays
html5lib
hurry.filesize
hyperlink
idna
incremental
Jinja2
Markdown
MarkupSafe
mysqlclient
numpy
oauthlib
pandas
phonenumbers
phonenumberslite
Pillow
pipenv
pycparser
PyHamcrest
Pyphen
PySocks
pysolr
python-dateutil
python-decouple
python3-openid
pytube
pytz
rcssmin
redis
reportlab
requests
requests-oauthlib
rjsmin
serious-django-permissions
six
soupsieve
South
sqlparse
tinycss2
tweepy
txaio
urllib3
virtualenv
virtualenv-clone
WeasyPrint
webencodings
wget
Whoosh
zope.interface
